+++ indispensable-cosef/src/commonTest/kotlin/at/asitplus/signum/indispensable/cosef/CoseSerializationTest.kt
@@ -4,7 +4,7 @@ import at.asitplus.signum.indispensable.CryptoSignature
 import at.asitplus.signum.indispensable.cosef.io.Base16Strict
 import at.asitplus.signum.indispensable.cosef.io.ByteStringWrapper
 import at.asitplus.signum.indispensable.cosef.io.coseCompliantSerializer
-
+import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.shouldBe
@@ -18,6 +18,18 @@ import kotlin.random.Random
 
 class CoseSerializationTest : FreeSpec({
 
+    "CoseSigned can not be constructed with ByteStringWrapper" {
+        val payload = ByteStringWrapper("StringType")
+        shouldThrow<IllegalArgumentException> {
+            CoseSigned<ByteStringWrapper<*>>(
+                protectedHeader = CoseHeader(algorithm = CoseAlgorithm.ES256),
+                unprotectedHeader = CoseHeader(),
+                payload = payload,
+                signature = CryptoSignature.RSAorHMAC(byteArrayOf())
+            )
+        }
+    }
+
     "Serialization is correct for byte array" {
         val payload = "This is the content.".encodeToByteArray()
         val cose = CoseSigned<ByteArray>(
@@ -26,26 +38,26 @@ class CoseSerializationTest : FreeSpec({
             payload = payload,
             signature = CryptoSignature.RSAorHMAC("bar".encodeToByteArray()) //RSAorHMAC because EC expects tuple
         )
-        val serialized = cose.serialize().encodeToString(Base16Strict).uppercase()
+        val serialized = cose.serialize(ByteArraySerializer()).encodeToString(Base16Strict).uppercase()
 
         serialized shouldContain "546869732069732074686520636F6E74656E742E" // "This is the content."
         serialized shouldContain "43A10126"
-        cose.getTypedPayload(ByteArraySerializer()).isFailure shouldBe true
+        cose.payload shouldBe payload
     }
 
     "Serialization is correct for data class" {
         val payload = DataClass("This is the content.")
-        val cose = CoseSigned.fromObject(
+        val cose = CoseSigned(
             protectedHeader = CoseHeader(algorithm = CoseAlgorithm.ES256),
             unprotectedHeader = CoseHeader(),
             payload = payload,
             signature = CryptoSignature.RSAorHMAC("bar".encodeToByteArray()) //RSAorHMAC because EC expects tuple
         )
-        val serialized = cose.serialize().encodeToString(Base16Strict).uppercase()
+        val serialized = cose.serialize(DataClass.serializer()).encodeToString(Base16Strict).uppercase()
 
         serialized shouldContain "546869732069732074686520636F6E74656E742E" // "This is the content."
         serialized shouldContain "43A10126"
-        cose.getTypedPayload(DataClass.serializer()).getOrThrow()?.value shouldBe payload
+        cose.payload shouldBe payload
     }
 
     "Serialize header" {
@@ -77,7 +89,8 @@ class CoseSerializationTest : FreeSpec({
                 "91aef0b0117e2af9a291aa32e14ab834dc56ed2a223444547e01f11d3b09" +
                 "16e5a4c345cacb36"
 
-        val cose = CoseSigned.deserialize(input.uppercase().decodeToByteArray(Base16Strict)).getOrThrow()
+        val cose = CoseSigned.deserialize(ByteArraySerializer(), input.uppercase().decodeToByteArray(Base16Strict))
+            .getOrThrow()
 
         cose.payload shouldBe "This is the content.".encodeToByteArray()
     }
@@ -95,6 +108,7 @@ class CoseSerializationTest : FreeSpec({
         val inputLibrary = CoseSigned.prepareCoseSignatureInput(
             protectedHeader = header,
             payload = payload,
+            serializer = ByteArraySerializer(),
         ).encodeToString(Base16())
 
         inputManual.shouldContain("Signature1".encodeToByteArray().encodeToString(Base16()))
@@ -114,6 +128,7 @@ class CoseSerializationTest : FreeSpec({
         val inputLibrary = CoseSigned.prepareCoseSignatureInput(
             protectedHeader = header,
             payload = payload,
+            serializer = DataClass.serializer(),
         ).encodeToString(Base16())
 
         inputManual.shouldContain("Signature1".encodeToByteArray().encodeToString(Base16()))