+++ indispensable-josef/src/jvmTest/kotlin/at/asitplus/signum/indispensable/josef/JwkTest.kt
@@ -44,7 +44,7 @@ class JwkTest : FreeSpec({
                 val own = cryptoPubKey.toJsonWebKey()
                 own.keyId shouldBe cryptoPubKey.jwkId
                 own.shouldNotBeNull()
-                println(own.serialize())
+
                 own.toCryptoPublicKey().getOrThrow().iosEncoded shouldBe cryptoPubKey.iosEncoded
                 CryptoPublicKey.fromDid(own.keyId!!) shouldBe cryptoPubKey
             }
@@ -65,7 +65,7 @@ class JwkTest : FreeSpec({
             certificateSha256Thumbprint = Random.nextBytes(32),
         )
 
-        val serialized = jwk.serialize().also { println(it) }
+        val serialized = jwk.serialize()
         val parsed = JsonWebKey.deserialize(serialized).getOrThrow()
 
         parsed shouldBe jwk
@@ -87,7 +87,7 @@ class JwkTest : FreeSpec({
                 certificateSha256Thumbprint = Random.nextBytes(32),
             )
 
-            val serialized = jwk.serialize().also { println(it) }
+            val serialized = jwk.serialize()
             val parsed = JsonWebKey.deserialize(serialized).getOrThrow()
 
             parsed shouldBe jwk